+++ src/permission-groups/dto/remove-agent-from-grop.dto.ts
@@ -1,6 +1,7 @@
+/* istanbul ignore file */
 import { OmitType } from '@nestjs/swagger';
 
-import { InviteAgentToGroupDto } from '~/assets/dto/invite-agent-to-group.dto';
+import { InviteAgentToGroupDto } from '~/permission-groups/dto/invite-agent-to-group.dto';
 
 export class RemoveAgentFromGroupDto extends OmitType(InviteAgentToGroupDto, [
   'permissions',
